split out htmlstring to a separate file
